Update README
Switch to org file for documentation. Previous documentation matched
the PyPI release, but not the current status of the code. Most of
changes correspond to the move from a command-line based workflow
towards Sylvia's REPL.

Added documentation on pronunciation inferencing, and explained rhyme
rules in more detail. Made it more clear that this program still
requires Python 2.