[emacs] Move towards an EPC backend
Communicating with SylviaConsole via Comint is a multiple degrees of
awkward, and it's the source of at least half the bugs I'm aware of so
far. Add some code to reimplement ELisp bindings using the EPC RPC
library (calling Python from ELisp). So far, it requires a fraction of
the effort and hasn't skipped a beat.