[rhyme] Restructure rhyme code
Remove PhoneticDictionary.getRhymes() and .getVowelmatches(), and add
.getRhymeRegex(). This function, which includes parts of the removed
APIs, generates a regex given a word and a rhyme level. We also now
use the "public" rhyme level names (loose, default, perfect)
internally.

NOTE! This changes the behavior of the "vowel matching" or "loose"
strategy. It now enforces that no new vowels can occur after the
matched sequence, which is more in keeping with the traditional
concept of a rhyme.

The removed functions can be replaced with:
  pd.regexSearch( pd.getRhymeRegex( "something", "perfect" ) )