[emacs] Simplify sylvia:echo-phonemes-at-point
Use the same logic as the -in-region version, which relies on the poem
APIs. This is more efficient, and reduces logic in the client.

This also has the pleasant side-affect that, if the word at point is
not in the dictionary, it's inferred pronunciation will be shown
(since Poem infers by default).